backend/template_generation_func/requirements.txt
functions-framework==3.*
google-cloud-storage==2.14.0
google-cloud-firestore==2.11.1
google-cloud-documentai==3.5.0
google-cloud-aiplatform==1.56.0
google-auth==2.29.0
PyPDF2==3.0.1
protobuf<4.0.0